requirements.txt
-i https://pypi.org/simple
aiodns==3.0.0
aiohttp==3.8.4 ; python_version >= '3.6'
aiosignal==1.3.1 ; python_version >= '3.7'
aiosqlite==0.18.0
async-timeout==4.0.2 ; python_version >= '3.6'
attrs==23.1.0 ; python_version >= '3.7'
beautifulsoup4==4.12.2
better-profanity==0.7.0
brotli==1.0.9
cffi==1.15.1
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
discord.py[speed]==2.2.3
frozenlist==1.3.3 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
multidict==6.0.4 ; python_version >= '3.7'
orjson==3.9.1
pillow==9.5.0
pycares==4.3.0
pycparser==2.21
soupsieve==2.4.1 ; python_version >= '3.7'
tzdata==2023.3
yarl==1.9.2 ; python_version >= '3.7'
